+++ R/internal_utility.R
@@ -1,4 +1,4 @@
-.validate_var = \(data, target, source, detrend = FALSE) {
+.validate_var = \(data, target, source = NULL, detrend = FALSE) {
   coords = NULL
   if (detrend) {
     if (inherits(data, "sf")) {
@@ -10,8 +10,10 @@
     }
     names(coords) = c("x", "y")
   }
-
-  var_indices = c(abs(target[1]), abs(source[1]))
+  
+  var_indices = abs(target[1])
+  if (!is.null(source)) var_indices = c(var_indices, abs(source[1]))
+  
   if (inherits(data, "sf")) {
     data = sf::st_drop_geometry(data[, var_indices, drop = FALSE])
   } else if (inherits(data, "SpatRaster")) {
@@ -19,7 +21,12 @@
   } else {
     data = data[, var_indices, drop = FALSE]
   }
-  names(data) = c("target", "source")
+
+  if (is.null(source)) {
+    names(data) = "target"
+  } else {
+    names(data) = c("target", "source")
+  }
 
   if (!all(apply(data, 2, typeof) %in% c("integer", "double"))) {
     stop("Non-numeric values detected in input data. 
@@ -29,11 +36,11 @@
   res = vector("list", 2)
   if (is.null(coords)) {
     res[[1]] = data[, 1, drop = TRUE]
-    res[[2]] = data[, 2, drop = TRUE]
+    if (!is.null(source)) res[[2]] = data[, 2, drop = TRUE]
   } else {
     data = as.data.frame(cbind(data, coords))
     res[[1]] = sdsfun::rm_lineartrend("target~x+y", data = data)
-    res[[2]] = sdsfun::rm_lineartrend("source~x+y", data = data)
+    if (!is.null(source)) res[[2]] = sdsfun::rm_lineartrend("source~x+y", data = data)
   }
 
   return(res)