+++ R/pc.R
@@ -55,32 +55,18 @@
 
 #' Pattern Causality
 #'
-#' @note `pc` only supports numeric input data.
-#'
-#' @param data Observation data.
-#' @param target Integer of column indice for the target variable.
+#' @inheritParams fnn
 #' @param source Integer of column indice for the source variable.
 #' @param libsizes (optional) Number of observations used.
-#' @param E (optional) Embedding dimensions.
 #' @param k (optional) Number of nearest neighbors used for projection.
-#' @param tau (optional) Step of lag.
-#' @param style (optional) Embedding style (`0` includes current state, `1` excludes it).
-#' @param lib (optional) Libraries indices.
-#' @param pred (optional) Predictions indices.
 #' @param boot (optional) Number of bootstraps to perform.
 #' @param random (optional) Whether to use random sampling.
 #' @param seed (optional) Random seed.
-#' @param dist.metric (optional) Distance measure to be used.
 #' @param zero.tolerance (optional) Maximum number of zeros tolerated in signature space.
 #' @param relative (optional) Whether to calculate relative changes in embedding.
 #' @param weighted (optional) Whether to weight causal strength.
-#' @param threads (optional) Number of threads used.
-#' @param higher.parallel (optional) Whether to use a higher level of parallelism.
 #' @param verbose (optional) Whether to show the progress bar.
 #' @param h (optional) Prediction horizon.
-#' @param detrend (optional) Whether to remove the linear trend.
-#' @param nb (optional) Neighbours list.
-#' @param ... Additional arguments to absorb unused inputs in method dispatch.
 #'
 #' @return A list (when `libsizes` is `NULL`) or data.frame.
 #' If `libsizes` is `NULL`, a list with two components is returned: